+++ frontend/src/metabase/status/components/DatabaseStatus/DatabaseStatus.tsx
@@ -1,11 +1,12 @@
 import React from "react";
+import { t } from "ttag";
+import Tooltip from "metabase/components/Tooltip";
 import useStatusVisibility from "../../hooks/use-status-visibility";
 import { Database } from "../../types";
 import {
   StatusRoot,
   StatusIconContainer,
   StatusIcon,
-  getIconName,
 } from "./DatabaseStatus.styled";
 
 interface Props {
@@ -21,15 +22,39 @@ const DatabaseStatus = ({ database }: Props) => {
   }
 
   return (
-    <StatusRoot status={database.initial_sync_status}>
-      <StatusIconContainer status={database.initial_sync_status}>
-        <StatusIcon
-          status={database.initial_sync_status}
-          name={getIconName(database)}
-        />
-      </StatusIconContainer>
-    </StatusRoot>
+    <Tooltip tooltip={getTooltip(database)}>
+      <StatusRoot status={database.initial_sync_status}>
+        <StatusIconContainer status={database.initial_sync_status}>
+          <StatusIcon
+            status={database.initial_sync_status}
+            name={getIconName(database)}
+          />
+        </StatusIconContainer>
+      </StatusRoot>
+    </Tooltip>
   );
+};
+
+const getTooltip = (database: Database) => {
+  switch (database.initial_sync_status) {
+    case "incomplete":
+      return t`Syncing ${database.name} …`;
+    case "complete":
+      return t`Done!`;
+    case "aborted":
+      return t`Error syncing`;
+  }
+};
+
+export const getIconName = (database: Database) => {
+  switch (database.initial_sync_status) {
+    case "incomplete":
+      return "database";
+    case "complete":
+      return "check";
+    case "aborted":
+      return "warning";
+  }
 };
 
 export default DatabaseStatus;
